# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de proveedores",
        "description": "Documentación de la API para gestión de proveedores y Productoos.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8080",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/lineas": {
            "get": {
                "tags": [
                    "Líneas"
                ],
                "summary": "Listar líneas",
                "operationId": "860ab3b4d96b6ac06f7bf5b6c8de1556",
                "responses": {
                    "200": {
                        "description": "Lista de líneas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Linea"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Líneas"
                ],
                "summary": "Crear línea",
                "operationId": "45c181950c7adcd5d2c2b5653e2718f8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Linea"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Línea creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Linea"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/lineas/{id}": {
            "get": {
                "tags": [
                    "Líneas"
                ],
                "summary": "Obtener línea por ID",
                "operationId": "b450e06ed534fbce34a7ace84466cf38",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Línea encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Linea"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Líneas"
                ],
                "summary": "Actualizar línea",
                "operationId": "b8201caed3697d0672cc28d5ece42771",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Linea"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Línea actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Linea"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Líneas"
                ],
                "summary": "Eliminar línea",
                "operationId": "479c3608cd21d62552432722f7a4b7f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Línea eliminada"
                    }
                }
            }
        },
        "/api/marcas": {
            "get": {
                "tags": [
                    "marcas"
                ],
                "summary": "Listar marcas",
                "operationId": "cdf1f15837825105957dc1b6581c9377",
                "responses": {
                    "200": {
                        "description": "Lista de marcas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Marca"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "marcas"
                ],
                "summary": "Crear marca",
                "operationId": "011cfb0f49c0cd34dc812d3c5e40e41f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Marca"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Marca creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Marca"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/marcas/{id}": {
            "get": {
                "tags": [
                    "marcas"
                ],
                "summary": "Obtener marca por ID",
                "operationId": "f044df503607a332772744171e7aebcc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marca encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Marca"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "marcas"
                ],
                "summary": "Actualizar marca",
                "operationId": "34d9e20c71578f8affd3a8a0c877d47c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Marca"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Marca actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Marca"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "marcas"
                ],
                "summary": "Eliminar marca",
                "operationId": "a11c924f91478a8599debeaa76b10428",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Marca eliminada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Linea": {
                "required": [
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Industrial"
                    },
                    "photo_url": {
                        "type": "string",
                        "example": "https://dominio.com/storage/lineas/industrial.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Marca": {
                "required": [
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Bosch"
                    },
                    "photo_url": {
                        "type": "string",
                        "example": "https://dominio.com/storage/marcas/bosch.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}